Upstream Patch: Fix wide character being incorrectly cleared in MODE_INSERT
Author: Shi Tian <shitian@cock.li>
Date:   Sun, 25 Jun 2023 05:38:33 +0000

Fix for wide character being incorrectly cleared on MODE_INSERT

Under insert mode, when inserting a normal character in front of
a wide character, the affected region is shifted to the right by
one cell. However, the empty cell is reset as if being a part of a
wide character, causing the following cell being mishandled as a
dummy cell.
To reproduce the bug:
	printf '\033[4h' # set MODE_INSERT
	printf 妳好
	printf '\033[4D'
	printf 'x'
	printf '\033[4l\n'